# ARROW-16117: [JS] Improve decode UTF8 performance