@domoritz 2022-04-04
Sweet. Thank you for the pull request. Can you open a Jira ticket and add it to the title? 

Also, we added memoization to dictionary vectors in the last release. If you have repeating strings, you can save a lot of time by not re-decoding strings.