@domoritz 2022-04-04
Arrow is pretty exciting for the browser and I think in the last version we made some pretty sweet API changes. There is definitely still lots of room for perf tricks and convenience methods.