@domoritz 2022-04-04
Thank you. Do you want to try running the benchmarks locally instead of waiting for the bot to get back to us?