@hzuo 2022-04-04
Looks like it's generally faster, but there's also a lot of variance in the two runs across other benchmarks which should not have been impacted, so unclear how much we can attribute to this PR:

https://gist.github.com/hzuo/4a39a661ddd33b740bb7bcb4d85aaa48