@domoritz 2022-04-04
Thanks for running these. If you rerun the strings tests only, and run them a few times, do you see a reliable difference?