@hzuo 2022-04-05
Is there a way via npm scripts or should I comment out cases in the benchmarking code?